Use hash_equals in password.php (and move it to funcs.php)
